refactor: Seperate AuthExceptionHandler
