+++ src/main/java/gift/exceptionhandler/GlobalExceptionHandler.java
@@ -1,7 +1,5 @@
 package gift.exceptionhandler;
 
-import gift.exception.FailedLoginException;
-import gift.exception.InvalidAccessTokenException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -13,18 +11,6 @@ public class GlobalExceptionHandler {
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(IllegalArgumentException.class)
     public String handleIllegalArgumentException(IllegalArgumentException exception) {
-        return exception.getMessage();
-    }
-
-    @ResponseStatus(HttpStatus.FORBIDDEN)
-    @ExceptionHandler(FailedLoginException.class)
-    public String handleFailedLoginException(FailedLoginException exception) {
-        return exception.getMessage();
-    }
-
-    @ResponseStatus(HttpStatus.FORBIDDEN)
-    @ExceptionHandler(InvalidAccessTokenException.class)
-    public String handleInvalidAccessTokenException(InvalidAccessTokenException exception) {
         return exception.getMessage();
     }
 }
